+++ terminaltexteffects/effects/effect_decrypt.py
@@ -63,7 +63,7 @@ class DecryptEffect:
         self.animating_chars: list[EffectCharacter] = []
         self.ciphertext_color = args.ciphertext_color
         self.plaintext_color = args.plaintext_color
-        self.character_discovered_gradient: graphics.Gradient = graphics.Gradient("ffffff", self.plaintext_color, 5)
+        self.character_discovered_gradient: graphics.Gradient = graphics.Gradient(["ffffff", self.plaintext_color], 5)
         self.encrypted_symbols: list[str] = []
         self.scenes: dict[str, graphics.Scene] = {}
         self.make_encrypted_symbols()