+++ CHANGELOG.md
@@ -14,7 +14,7 @@
  * Paths have replaced Waypoints as the motion target specification object. Paths group Waypoints together and allow for easing
    motion and animations across an arbitrary number of Waypoints. Single Waypoint Paths are supported and function the same as
    Waypoints did previously. Paths can be looped with the loop argument. 
- * Quadratic bezier curves are supported. Control points are specified in the Waypoint object signature. When a control point
+ * Quadratic and Cubic bezier curves are supported. Control points are specified in the Waypoint object signature. When a control point
    is specified, motion will be curved from the prior Waypoint to the Waypoint with the control point, using the control point
    to determine the curve. Curves are supported within Paths.
  * New EventHandler.Event PATH_HOLDING is triggered when a Path enters the holding state.
@@ -31,9 +31,13 @@
  * graphics.Animation.random_color() is now a static method.
  * motion.Motion.find_coords_in_circle() now generates 7*radius coords in each inner-circle.
  * BlackholeEffect uses chain_paths() and benefits from better circle support for a much improved blackhole animation.
+ * BlackholeEffect singularity Paths are curved towards center lines.
  * EventHandler.Event.WAYPOINT_REACHED removed and split into two events, PATH_HOLDING and PATH_COMPLETE.
  * EventHandler.Event.PATH_COMPLETE is triggered when the final Path Waypoint is reached AND holding time reaches 0.
  * Fireworks effect uses Paths and curves to create a more realistic firework explosion.
+ * Crumble effect uses control points to create a curved vacuuming phase.
+ * graphics.Gradient accepts an arbitrary number of color stops. The number of steps applies between each color stop.
+ * motion.find_coords_in_circle() and motion.find_coords_in_rect() no longer take a num_points argument. All points in the area are returned.
 
 ### Bug Fixes
  * Fixed looping animations when synced to Path not resetting properly.